server/requirements.txt

Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-framework==0.1.0
djoser==2.1.0
django-phonenumber-field==6.0.0
django-weekday-field == 1.1.0

